Some arrangements to use the newly created mdp
